+++ packages/server-runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cosmicdrift/kumiko-server-runtime
 
+## 0.270.0
+
+### Patch Changes
+
+- Updated dependencies [dba0a60]
+  - @cosmicdrift/kumiko-framework@0.270.0
+  - @cosmicdrift/kumiko-bundled-features@0.270.0
+
 ## 0.269.2
 
 ### Patch Changes